pr comment for workspace model
