+++ src/app/core/models/db/workspaces.model.ts
@@ -1,5 +1,3 @@
-import { IntacctOnboardingState } from "../enum/enum.model";
-
 export type Workspace = {
   id: number;
   name: string;
@@ -10,11 +8,4 @@ export type Workspace = {
   updated_at: Date;
 }
 
-export interface IntacctWorkspace extends Workspace {
-  last_synced_at: Date;
-  ccc_last_synced_at: Date;
-  source_synced_at: Date;
-  cluster_domain: string;
-  destination_synced_at: Date;
-  onboarding_state: IntacctOnboardingState;
-}
+
