+++ src/app/core/models/si/db/workspaces.model.ts
@@ -1,0 +1,11 @@
+import { Workspace } from "../../db/workspaces.model";
+import { IntacctOnboardingState } from "../../enum/enum.model";
+
+export interface IntacctWorkspace extends Workspace {
+    last_synced_at: Date;
+    ccc_last_synced_at: Date;
+    source_synced_at: Date;
+    cluster_domain: string;
+    destination_synced_at: Date;
+    onboarding_state: IntacctOnboardingState;
+  }
